Using Android Rotation Vector Success
Updated historyLength to 10,000 rather than 15. Judy had previously been
trying to eliminate noise in the signal by averaging over 15 samples. I
changed this to be a low pass filter with a buffer equal to the
historyLength, so making this a fair bit longer made sense. This hasn't
been tuned too precisely, but just eyeballing the low pass effects on
the theta and thetaDot variables through the bash plotting. I just made
sure the low pass was filtering some amount of noise, while still
maintaining a high enough speed to react in real time with the tilt of
the phone. Current settings should likely work under any conditions the
Hubee wheels can provide, but if faster/more powerful motors ever are
used, this may need to be updated.