Python PID Control Working but only over bluetooth
As IOIO USB requires USB debugging to be off, I cannot send python adb code to the phone. As a workaround I can connect the phone to the IOIO board via bluetooth and keep USB debugging on. The lag is noticable though and cannot control the bot well enough to balance.